WIP - improve tests Extract TestScenario & TestScenarioBuilder
