+++ exercise/kotlin/day13/src/test/kotlin/santamarket/model/TestScenario.kt
@@ -1,0 +1,7 @@
+package santamarket.model
+
+class TestScenario(val catalog: FakeCatalog, val sleigh: ShoppingSleigh, val elf: ChristmasElf) {
+    fun checkout(): Receipt {
+        return elf.checksOutArticlesFrom(sleigh)
+    }
+}
